+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+## [0.5.0] - 2020-05-28
+
 ### Added
 
 - New function `Flop.validate!/2`.